+++ packages/material-ui-system/src/styleFunctionSx/extendSxProp.js
@@ -1,0 +1,31 @@
+import { deepmerge } from '@material-ui/utils';
+import { propToStyleFunction } from '../getThemeValue';
+
+const splitProps = (props) => {
+  const result = {
+    systemProps: {},
+    otherProps: {},
+  };
+
+  Object.keys(props).forEach((prop) => {
+    if (propToStyleFunction[prop]) {
+      result.systemProps[prop] = props[prop];
+    } else {
+      result.otherProps[prop] = props[prop];
+    }
+  });
+
+  return result;
+};
+
+export default function extendSxProp(props) {
+  const { sx: inSx , ...other } = props;
+  const { systemProps, otherProps } = splitProps(other);
+
+  const sx = systemProps ? deepmerge(systemProps, inSx) : inSx;
+
+  return {
+    ...otherProps,
+    sx,
+  };
+}
