+++ packages/material-ui-system/src/styleFunctionSx/index.js
@@ -1,1 +1,2 @@
 export { default } from './styleFunctionSx';
+export { default as extendSxProp } from './extendSxProp';
